requirements.txt
apache-airflow-providers-amazon==6.0.0
astronomer-providers[amazon]==1.11.1